Use more of std::string and fmt::sprintf.
Also, investigated the `%p` issue. Explicit casts to `(void *)` can still work. Unsure if it's a good idea all the time, but one step at a time perhaps.

The only usages of `vssprintf` now are tied to our custom `ssprintf` function. I do not trust that just replacing the functionality with `cppformat` will work here. That is for later tonight I guess.