Compile the Texture Font Generator with cmake.
This program is not built by default. Enable one of the following options to utilize it:

* WITH_FULL_RELEASE
* WITH_TEXTURE_GENERATOR

It is heavily advised to make sure the MFC library is installed for Visual Studio and the computer is restarted afterwards.

As a bonus, `zlib1.dll` can go away now.

The old visual studio projects are still in play in case others still struggle with the code fixes.