Create a new Pump skin for NewField.
This noteskin allows both classic and rhythm style gameplay through the skin parameters. Defaults and other choices are as follows:

* `color_style`: `Classic` (also `Rhythm`)
* `rhythm_quarter_color`: `Red` (also `Blue`)
* `frame_one_color`: `Beat` (also `Target`)
* `frame_cycle_beats`: `2` (also `1`)

SVG files are included since they are the base files used to generate the PNG files. Future editors: remember to modify the SVG and export to PNG later.

There is one primary limitation at this time. There are no lifts. Currently, they are mapped to mines.

More changes may come in later, but this should be a good start.