Partially fix song repetition in generated courses.
The old algorithm for generating courses has a tendency to repeat songs.
This commit prevents that from happening. Songs will instead shuffle, and will
only repeat if the current song list has finished.

This is a partial fix, since for some reason, the method responsible
for getting all the songs in a song group is not doing its work properly.

That means for a course that's supposed to have a 100 songs, only 95 of them
would appear. The other 5 will never appear, and the last 5 will be a repeat
of the first 5 songs.