XInput support for the DirectInput driver (#1440)
* XInput support for the DirectInput driver

The XBox360 controller driver for windows has a flaw where the DirectInput backwards-compatible driver registers dance pad arrows as a hat and will not allow opposite arrows to be pressed at the same time. This is not a problem in the XInput interface since the arrows are registered as a DPad and each arrow is a button.

This patch adds XInput support for the DirectInput driver. On initialization XInput controllers will be detected and registered first, followed by enumerating DirectInput devices. During DirectInput enumeration any XInput-capable controllers will be bypassed so that they aren't registered twice.

* Fix joystick input for xinput driver

Needed to compensate for deadzones, magnitude of the inputs, and that the Y axis was reversed.