Another Amends, with serious proposal
The reason I'm doing this pull requests is to propose the devleopers to accept pull requests #1224 (Updated Korean Glyphs) at your earliest convenience in order to show the Korean Translation properly at Stepmania. I want to stress out fact that the current version of Stepmania fails to show almost every character that are included in this file - which also denotes that ko.ini in _fallback and default - , since the Korean letters provided here cover onlt 2~3% of the 11,172 letters of the Korean Alphabet(Hangul.)

For more information, please read the issue #1122 by author of the pull request #1124.