backport legacy credits from master (#1906)
* Restore lost Stepmania Team credits

Back in 2009, an overhaul was made to the theme to replace the credits
screen. That new screen had a short list of credits, along with a note to
"remake this list". That never happened, and over time, the original credits
list was lost in the shuffle. People that worked on Stepmania in the past,
who felt their contributions were enough to warrant credit, should still
get credit.

This list was pulled directly from the git history under hash
fb1165e6be1cd1e40193b7d6a100c2525a670e04
It could probably be cleaned up a bit. (For instance, the theme has been
completely redone. Should we still include the Graphics and Theme sections?
Also, the web section. And the thanks section? Those are pretty much just
shoutouts.)

* Name change. Deal with it >8)

* Change encoding of restored credits to utf-8

The list used "fancy" quotes. I should have saved this at utf-8 to
begin with.