add 21:9 and 9:16 to ConfAspectRatio in _fallback
This adds choices for 21:9 (Ultrawide) and 9:16 (widescreen display in
portrait mode) to the _fallback theme.  Previously, users would have to
manually edit Preferences.ini to achieve either of these.

Since "21:9" is the common marketing term for what is actually 64:27,
I've slightly extended GenChoices() in ConfAspectRatio() to use a "name"
string from the winFracs table is one is provided. If not, the original
behavior of generating user-facing choices from the raw values is
preserved.

See: https://wikipedia.org/wiki/21:9_aspect_ratio

I updated the global AspectRatios table in _fallback/02 Utilities.lua to
match.