Change the default binding for P2/back from hyphen to backslash
This is a workaround to make the hyphen key usable in text entry
screens. It allows to add a hyphen to profile names and set a negative
offset in edit mode.

Technically backslash is also a typeable character, but it seems to be
less common names in user names and not necessary for typing negative
offsets.

Fwiw, in outfox P2/back is unbound.