macos: Workaround for include conflict
New version of c++ std lib has an include file called `version`.

Setting the include path to the root of a dependency that has a file called `VERSION`,
makes it try to include that file instead, and it seems some other stdlib files try to include it.
Additional macos by default is case-insensitive.

This just removes the `VERSION` file from the vendored deps,
and also attempts to patch the ffmpeg dependency