Fix macos (#2216)
* macos: Fix zlib compilation

* macos: Fix tomcrypt compilation

it seems HMAC is also needed

* macos: Add LightsDriver_Export to build

This seemed to be required, and at least compiled fine.

* macos: Workaround for include conflict

New version of c++ std lib has an include file called `version`.

Setting the include path to the root of a dependency that has a file called `VERSION`,
makes it try to include that file instead, and it seems some other stdlib files try to include it.
Additional macos by default is case-insensitive.

This just removes the `VERSION` file from the vendored deps,
and also attempts to patch the ffmpeg dependency

* macos: Workaround a keyhandling crash

When pressing a button with a mat/joypad(?), it ends up here with keyCode zero,
and then for some reason TISGetInputSourceProperty crashes, which I believe to be
a system bug. Worked around by not handling zero keyCode.